backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pymongo
mongoengine
django-cors-headers